added socket packets

+++ REghZyIOWrapperV2/Connections/Serial/SerialDataStream.cs
@@ -2,12 +2,15 @@
 using REghZyIOWrapperV2.Streams;
 
 namespace REghZyIOWrapperV2.Connections.Serial {
+    /// <summary>
+    /// A data stream that uses the <see cref="SerialPort.BaseStream"/> for reading and writing data
+    /// </summary>
     public class SerialDataStream : DataStream {
         private readonly SerialPort port;
 
         public SerialPort Port { get => this.port; }
 
-        public override int BytesAvailable => this.port.BytesToRead;
+        public override long BytesAvailable => this.port.BytesToRead;
 
         public SerialDataStream(SerialPort port) : base(port.BaseStream) {
             this.port = port;
